+++ R/are_lgl_ish.R
@@ -71,8 +71,7 @@ are_lgl_ish.numeric <- function(x, ...) {
 
 #' @export
 are_lgl_ish.character <- function(x, ...) {
-  cast <- as.logical(toupper(x))
-  !xor(is.na(x), is.na(cast)) | are_dbl_ish(x, ...)
+  .Call(ffi_chr_to_lgl, x)[["valid"]]
 }
 
 #' @export